refactor: merge render body into document
