+++ crates/publish/src/render.rs
@@ -1,4 +1,3 @@
-mod body;
 mod bookmark;
 mod document;
 mod html;
